+++ hello/templates/index.html
@@ -13,7 +13,7 @@
           <input
               type="text"
               id="input_text"
-              name="full_name"
+              name="Full Name"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Full Name (optional)"
           />
@@ -26,7 +26,7 @@
           <input
               type="email"
               id="input_text"
-              name="email_address"
+              name="Email Address"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Email Address (optional)"
           />
@@ -39,7 +39,7 @@
           <input
               type="text"
               id="input_text"
-              name="sex_assigned_at_birth"
+              name="Sex"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder=""
           />
@@ -52,7 +52,7 @@
           <input
               type="text"
               id="input_text"
-              name="current_job_title_and_company"
+              name="Current Job Title and Company"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Current Job Title and Company"
           />
@@ -65,7 +65,7 @@
           <input
               type="text"
               id="input_text"
-              name="industry"
+              name="Industry"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Industry"
           />
@@ -78,7 +78,7 @@
           <input
               type="text"
               id="input_text"
-              name="working_region"
+              name="Working Region"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Working Region"
           />
@@ -90,7 +90,7 @@
           </label>
           <textarea
               id="input_textarea"
-              name="productivity_limitations"
+              name="Productivity Limitations"
               rows="4"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Productivity Limitations"
@@ -103,7 +103,7 @@
           </label>
           <textarea
               id="input_textarea"
-              name="career_goals"
+              name="Career Goals"
               rows="4"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Career Goals"
@@ -117,7 +117,7 @@
           <input
               type="text"
               id="input_text"
-              name="favorite_scent"
+              name="Favorite Scent"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Favorite Scent (optional)"
           />
@@ -129,7 +129,7 @@
           </label>
           <textarea
               id="input_textarea"
-              name="positive_childhood_scent_association"
+              name="Positive Childhood Scent Association"
               rows="4"
               class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition duration-300 ease-in-out"
               placeholder="Positive Childhood Scent Association"
